Extract JSGLR2Variant.ParserVariant to upper-level class
