+++ org.spoofax.jsglr2.integration/src/main/java/org/spoofax/jsglr2/integration/IntegrationVariant.java
@@ -9,6 +9,7 @@ import org.spoofax.jsglr2.JSGLR2Variants;
 import org.spoofax.jsglr2.imploder.ImploderVariant;
 import org.spoofax.jsglr2.parseforest.ParseForestConstruction;
 import org.spoofax.jsglr2.parseforest.ParseForestRepresentation;
+import org.spoofax.jsglr2.parser.ParserVariant;
 import org.spoofax.jsglr2.reducing.Reducing;
 import org.spoofax.jsglr2.stack.StackRepresentation;
 import org.spoofax.jsglr2.stack.collections.ActiveStacksRepresentation;
@@ -18,11 +19,11 @@ import org.spoofax.jsglr2.tokens.TokenizerVariant;
 public class IntegrationVariant {
     public final ParseTableVariant parseTable;
     public final JSGLR2Variants.Variant jsglr2;
-    public final JSGLR2Variants.ParserVariant parser;
+    public final ParserVariant parser;
     public final ImploderVariant imploder;
     public final TokenizerVariant tokenizer;
 
-    public IntegrationVariant(ParseTableVariant parseTableVariant, JSGLR2Variants.ParserVariant parserVariant,
+    public IntegrationVariant(ParseTableVariant parseTableVariant, ParserVariant parserVariant,
         ImploderVariant imploderVariant, TokenizerVariant tokenizer) {
         this.parseTable = parseTableVariant;
         this.jsglr2 = new JSGLR2Variants.Variant(parserVariant, imploderVariant, tokenizer);
@@ -59,17 +60,17 @@ public class IntegrationVariant {
     public static List<IntegrationVariant> testVariants() {
         //@formatter:off
         return Arrays.asList(
-            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.Basic,  ParseForestConstruction.Full, StackRepresentation.Basic,          Reducing.Basic),       ImploderVariant.TokenizedRecursive,   TokenizerVariant.Null),
-            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.DisjointSorted, ProductionToGotoRepresentation.JavaHashMap), new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.Basic,  ParseForestConstruction.Full, StackRepresentation.Basic,          Reducing.Basic),       ImploderVariant.Recursive,            TokenizerVariant.Recursive),
-            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Hybrid, ParseForestConstruction.Full, StackRepresentation.Hybrid,         Reducing.Basic),       ImploderVariant.Iterative,            TokenizerVariant.Iterative),
-            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Hybrid, ParseForestConstruction.Full, StackRepresentation.HybridElkhound, Reducing.Elkhound),    ImploderVariant.TokenizedRecursive,   TokenizerVariant.Null),/*
+            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.Basic,  ParseForestConstruction.Full, StackRepresentation.Basic,          Reducing.Basic),       ImploderVariant.TokenizedRecursive,   TokenizerVariant.Null),
+            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.DisjointSorted, ProductionToGotoRepresentation.JavaHashMap), new ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.Basic,  ParseForestConstruction.Full, StackRepresentation.Basic,          Reducing.Basic),       ImploderVariant.Recursive,            TokenizerVariant.Recursive),
+            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Hybrid, ParseForestConstruction.Full, StackRepresentation.Hybrid,         Reducing.Basic),       ImploderVariant.Iterative,            TokenizerVariant.Iterative),
+            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Hybrid, ParseForestConstruction.Full, StackRepresentation.HybridElkhound, Reducing.Elkhound),    ImploderVariant.TokenizedRecursive,   TokenizerVariant.Null),/*
             new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Hybrid, ParseForestConstruction.Optimized, StackRepresentation.Hybrid,         Reducing.Basic)),
             new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Hybrid, ParseForestConstruction.Optimized, StackRepresentation.HybridElkhound, Reducing.Elkhound)),*/
-            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Incremental, ParseForestConstruction.Full, StackRepresentation.Hybrid,    Reducing.Incremental), ImploderVariant.RecursiveIncremental, TokenizerVariant.Recursive),
+            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.LinkedHashMap, ForActorStacksRepresentation.LinkedHashMap, ParseForestRepresentation.Incremental, ParseForestConstruction.Full, StackRepresentation.Hybrid,    Reducing.Incremental), ImploderVariant.RecursiveIncremental, TokenizerVariant.Recursive),
             // data-dependent parser
-            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.DataDependent,   ParseForestConstruction.Full, StackRepresentation.Basic, Reducing.DataDependent),   ImploderVariant.TokenizedRecursive, TokenizerVariant.Null),
+            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.DataDependent,   ParseForestConstruction.Full, StackRepresentation.Basic, Reducing.DataDependent),   ImploderVariant.TokenizedRecursive, TokenizerVariant.Null),
             // layout-sensitive parser
-            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.LayoutSensitive, ParseForestConstruction.Full, StackRepresentation.Basic, Reducing.LayoutSensitive), ImploderVariant.TokenizedRecursive, TokenizerVariant.Null)
+            new IntegrationVariant(new ParseTableVariant(ActionsForCharacterRepresentation.Separated,      ProductionToGotoRepresentation.ForLoop),     new ParserVariant(ActiveStacksRepresentation.ArrayList,     ForActorStacksRepresentation.ArrayDeque,    ParseForestRepresentation.LayoutSensitive, ParseForestConstruction.Full, StackRepresentation.Basic, Reducing.LayoutSensitive), ImploderVariant.TokenizedRecursive, TokenizerVariant.Null)
         );
         //@formatter:on
     }
