+++ org.spoofax.jsglr2/src/main/java/org/spoofax/jsglr2/reducing/ReduceManagerFactory.java
@@ -1,7 +1,6 @@
 package org.spoofax.jsglr2.reducing;
 
 import org.metaborg.parsetable.IParseTable;
-import org.spoofax.jsglr2.JSGLR2Variants;
 import org.spoofax.jsglr2.datadependent.DataDependentReduceManager;
 import org.spoofax.jsglr2.elkhound.AbstractElkhoundStackNode;
 import org.spoofax.jsglr2.elkhound.ElkhoundReduceManager;
@@ -16,6 +15,7 @@ import org.spoofax.jsglr2.parseforest.IParseForest;
 import org.spoofax.jsglr2.parseforest.ParseForestManager;
 import org.spoofax.jsglr2.parser.AbstractParse;
 import org.spoofax.jsglr2.parser.IParseState;
+import org.spoofax.jsglr2.parser.ParserVariant;
 import org.spoofax.jsglr2.stack.AbstractStackManager;
 import org.spoofax.jsglr2.stack.IStackNode;
 
@@ -52,7 +52,7 @@ public interface ReduceManagerFactory
         StackManager extends AbstractStackManager<ParseForest, StackNode, ParseState, Parse>>
     //@formatter:on
     ReduceManagerFactory<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse, StackManager, org.spoofax.jsglr2.reducing.ReduceManager<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse>>
-        reduceManagerFactory(JSGLR2Variants.ParserVariant parserVariant) {
+        reduceManagerFactory(ParserVariant parserVariant) {
         return (parseTable, stackManager, parseForestManager) -> new org.spoofax.jsglr2.reducing.ReduceManager<>(
             parseTable, stackManager, parseForestManager, parserVariant.parseForestConstruction);
     }
@@ -68,7 +68,7 @@ public interface ReduceManagerFactory
         StackManager extends ElkhoundStackManager<ParseForest, StackNode, ParseState, Parse>>
     //@formatter:on
     ReduceManagerFactory<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse, StackManager, ElkhoundReduceManager<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse>>
-        elkhoundReduceManagerFactory(JSGLR2Variants.ParserVariant parserVariant) {
+        elkhoundReduceManagerFactory(ParserVariant parserVariant) {
         return (parseTable, stackManager, parseForestManager) -> new ElkhoundReduceManager<>(parseTable, stackManager,
             parseForestManager, parserVariant.parseForestConstruction);
     }
@@ -84,7 +84,7 @@ public interface ReduceManagerFactory
         StackManager extends AbstractStackManager<ParseForest, StackNode, ParseState, Parse>>
     //@formatter:on
     ReduceManagerFactory<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse, StackManager, DataDependentReduceManager<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse>>
-        dataDependentReduceManagerFactory(JSGLR2Variants.ParserVariant parserVariant) {
+        dataDependentReduceManagerFactory(ParserVariant parserVariant) {
         return (parseTable, stackManager, parseForestManager) -> new DataDependentReduceManager<>(parseTable,
             stackManager, parseForestManager, parserVariant.parseForestConstruction);
     }
@@ -100,7 +100,7 @@ public interface ReduceManagerFactory
         StackManager extends AbstractStackManager<ParseForest, StackNode, ParseState, Parse>>
     //@formatter:on
     ReduceManagerFactory<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse, StackManager, LayoutSensitiveReduceManager<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse>>
-        layoutSensitiveReduceManagerFactory(JSGLR2Variants.ParserVariant parserVariant) {
+        layoutSensitiveReduceManagerFactory(ParserVariant parserVariant) {
         return (parseTable, stackManager, parseForestManager) -> new LayoutSensitiveReduceManager<>(parseTable,
             stackManager, parseForestManager, parserVariant.parseForestConstruction);
     }
@@ -116,7 +116,7 @@ public interface ReduceManagerFactory
         StackManager extends AbstractStackManager<ParseForest, StackNode, ParseState, Parse>>
     //@formatter:on
     ReduceManagerFactory<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse, StackManager, IncrementalReduceManager<ParseForest, ParseNode, Derivation, StackNode, ParseState, Parse>>
-        incrementalReduceManagerFactory(JSGLR2Variants.ParserVariant parserVariant) {
+        incrementalReduceManagerFactory(ParserVariant parserVariant) {
         return (parseTable, stackManager, parseForestManager) -> new IncrementalReduceManager<>(parseTable,
             stackManager, parseForestManager, parserVariant.parseForestConstruction);
     }
