+++ org.spoofax.jsglr2/src/test/java/org/spoofax/jsglr2/incremental/diff/ProcessUpdatesTest.java
@@ -3,7 +3,6 @@ package org.spoofax.jsglr2.incremental.diff;
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
-import org.spoofax.jsglr2.JSGLR2Variants;
 import org.spoofax.jsglr2.incremental.EditorUpdate;
 import org.spoofax.jsglr2.incremental.IncrementalParse;
 import org.spoofax.jsglr2.incremental.parseforest.IncrementalCharacterNode;
@@ -11,6 +10,7 @@ import org.spoofax.jsglr2.incremental.parseforest.IncrementalParseForest;
 import org.spoofax.jsglr2.incremental.parseforest.IncrementalParseNode;
 import org.spoofax.jsglr2.parseforest.ParseForestConstruction;
 import org.spoofax.jsglr2.parseforest.ParseForestRepresentation;
+import org.spoofax.jsglr2.parser.ParserVariant;
 import org.spoofax.jsglr2.parser.observing.ParserObserving;
 import org.spoofax.jsglr2.reducing.Reducing;
 import org.spoofax.jsglr2.stack.StackRepresentation;
@@ -20,11 +20,10 @@ import org.spoofax.jsglr2.stack.collections.ForActorStacksRepresentation;
 public class ProcessUpdatesTest {
 
     private final ProcessUpdates<?> processUpdates =
-        new ProcessUpdates<>(new IncrementalParse<>(
-            new JSGLR2Variants.ParserVariant(ActiveStacksRepresentation.standard(),
-                ForActorStacksRepresentation.standard(), ParseForestRepresentation.standard(),
-                ParseForestConstruction.standard(), StackRepresentation.standard(), Reducing.standard()),
-            "", "", new ParserObserving<>()));
+        new ProcessUpdates<>(new IncrementalParse<>(new ParserVariant(ActiveStacksRepresentation.standard(),
+            ForActorStacksRepresentation.standard(), ParseForestRepresentation.standard(),
+            ParseForestConstruction.standard(), StackRepresentation.standard(), Reducing.standard()), "", "",
+            new ParserObserving<>()));
 
     @Test public void testDeleteSubtree() {
         IncrementalParseNode previous = node(node(0, 1), node(2, 3), node(4, 5));
